Split HTTP scaffold tests
Move health endpoint, lifespan hook, and runtime environment wiring
coverage into focused test modules with shared scaffold support helpers.
Remove the file-level `too-many-lines` suppression from the original test.
